fix(dap): `loop` error when auto-detecting mason.nvim codelldb package
